create Review component and iclude it in Profile
