+++ src/pages/TeacherDetails/index.js
@@ -4,15 +4,10 @@ import { useDispatch, useSelector } from "react-redux";
 import { fetchTeacherDetails, fetchTeachers } from "../../store/teacher/actions";
 import { fetchSessions } from "../../store/session/actions";
 import Profile from "../../components/Profile";
-import PostReview from "../../components/PostReview";
 import { selectTeacherById } from "../../store/teacher/selectors";
-import { selectUser } from "../../store/user/selectors";
-import { Button } from "react-bootstrap";
 
 export default function TeacherDetails() {
-  const [showForm, setShowForm] = useState(false);
 
-  const { token } = useSelector(selectUser);
   const { id } = useParams();
   const dispatch = useDispatch();
 
@@ -32,12 +27,6 @@ export default function TeacherDetails() {
 
   return (
     <div>
-      {!showForm && token &&(
-        <Button onClick={() => setShowForm(!showForm)}>Make a review</Button>
-      )}
-      {showForm && (
-        <PostReview hideForm={() => setShowForm(!showForm)} id={id} />
-      )}
       {teacherDetails && <Profile profile={teacherDetails} teacher/>}
     </div>
   );
